Save Quaderno Contact IDs associated to the billing name
Only create a new contact in Quaderno when the billing name had
never been used before. Not everytime the billing name changes.

The plugin now stores in the user metadata the different billing names associated
to the Quaderno Contacts that were created when using those names.

This way, if a buyer changes her billing name back and forward, a new
contact will not be created for every change.